wb1/requirements.txt
autopep8==1.5.7
boto3==1.19.2
botocore==1.22.2
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
jmespath==0.10.0
pycodestyle==2.8.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
